Refactor company action tests around service behavior
